backend/requirements.txt
# FocusAI Backend Dependencies (MVP 简化版)

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# HTTP Client (for API calls and crawling)
httpx>=0.24.0
aiohttp>=3.8.0

# HTML Parsing (for news crawling)
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OpenAI-compatible API client (for SiliconFlow/DeepSeek)
openai>=1.0.0

# Scheduling (for auto-crawling)
apscheduler>=3.10.0

# Web Search (for tools/cases recommendation)
tavily-python>=0.3.0

# Utilities
python-dateutil>=2.8.0
